Added compression option
This introduces optional snappy compression to the configuration.

Why snappy?

Typically, snappy represents extremely low CPU overhead but still gives
worthwhile compression (~50% in my unscientific testing with some real
life protocol buffers messages)

Alternatives to snappy include lz4, but the advantage of snappy,
especially in Go is the existence of optimised asm code, making it ~10x
faster than lz4.

Why configuration?

Given it's characteristics, there is an argument to just always use
snappy compression and not add the complexity of another config option,
but there are two arguments against this:

1) The uncompressed version is already being used and this would break
compatability
2) It's possibly to imagine cases in future of large, low throughput
messages where other, slower and better compressing algorithms might be
wanted.

The default compression is None (for compatability), and for the moment
CompressionTypeSnappy is the only supported one.

This could be implemented instead as a decorator StreamStore in straw if
we can find a reasonable way to implement Stat().Size() correctly.

Question : Should we support something like CompressionTypeAutoDetect on
the consume side and haave it work on a per-file basis?  That would allow
flexibility of switching on compression for future files without a
migration effort.   If we want this, we might want to do it before
merging this.