Explicit Flush
Sometimes, callers need to know that what they have written so far is
safe and this is not guaranteed for the underlying straw stream until
Close is called, which only happens when we flush.

This introduces the ability to Flush() explicity, in addition to the
existing time and message count based approaches.