Handle partially written files
If we see EOF earlier than we expect it, assume the file is still being
written to, which is the most common case of this situation.  Note that
this can not happen for all types of straw backend.  In particular, s3
can never hit this case, but the filesystem case can.

We may in future want to allow the consumer to specifiy a max-retry type
parameter to bail out eventually, assuming that there is a corrupted
file.  If we ever see that in the wild, we can fix it then.