Fix memory leak
This moves the defer, that is intended to ensure the last input is closed
regardless of how we exit, to outside the main loop.  Being inside was
not only wrong (we should only check the last one, not every one) but it
caused a lingering reference to every stream opened, causing excessive
memory usage.