Change license to MIT
Since this is a library, and library code, when statically linked as it
is with Go, effectively becomes strong copyleft, whereas we wanted weak
copyleft.   Instead, use MIT.

In general, we should default to MIT for Go libraries that are public
and supported.

This change of license at this time is fine because all authors so far
are Utility Warehouse people.