Prevent zero length messages
Due to the design of the on-disk format of freezer, it does not support
zero length messages.  If we allow a zero length message to be written
by the sink, it is indistinguishable om disk from the internally used
end marker, and when being read as a source, is detected as a corrupt
stream.

Avoid this problem by explicitly dis-allowing empty messages and
returning an error if the caller attempts to write one.